refactor(nixvim): fix cmp select item behavior
This commit refactors the `cmp.mapping.select_next_item` behavior to only perform a selection when the user presses `Tab`.

This change fixes an issue where the `cmp.mapping.select_next_item` behavior would perform a selection when the user pressed `Enter`. This could lead to unexpected behavior when the user wanted to expand a snippet or confirm a selection.